[#84] refactor: utils 함수로 refactoring
